- removed some "using std::cerr" etc from .h or .inl files to avoid conflicts. This then needed changes in some files that include these files - a few extra explicit float or int casts to reduce number of warnings by VC++ - replaced a few output lines to std::cerr with calls to stir::info

+++ src/include/stir/ProjDataFromStream.inl
@@ -56,7 +56,7 @@ ProjDataFromStream::find_segment_index_in_sequence(const int segment_num) const
 #endif
   // TODO do some proper error handling here
   assert(iter !=  segment_sequence.end());
-  return iter - segment_sequence.begin();
+  return static_cast<int>(iter - segment_sequence.begin());
 }
